now it won't buy a trillion shares every time :smiley:

+++ main.py
@@ -1,5 +1,5 @@
 from yahoo_finance import Share
-import time, json, sys
+import time, json, sys, datetime
 
 class Wallet():
 
@@ -10,36 +10,52 @@ class Wallet():
         obj = ShareObj(ID)
         obj.refresh()
 
-        self.cash -= float(obj.getPrice())
+        if(self.cash > float(obj.getPrice())): #if you have sufficient funds
+            self.cash -= float(obj.getPrice()) #buy that shit
 
-        with open('data.json', 'r+') as file:
-            try:
-                data = json.load(file) # read file
-            except json.decoder.JSONDecodeError:
-                print('check your json file')
-                sys.exit(0)
+            with open('data.json', 'r+') as file:
+                try:
+                    data = json.load(file) # read file
+                except json.decoder.JSONDecodeError:
+                    print('Error loading JSON - Check your JSON file')
+                    sys.exit(0)
 
-            # wipe file
-            file.seek(0)
-            file.truncate()
+                # wipe file
+                file.seek(0)
+                file.truncate()
 
-            try:
-                data['shares'].append({ # add new price to appropriate id
-                    "id": obj.getID(),
-                    "price": obj.getPrice()
-                })
-            except AttributeError:
-                # uh oh!!! that ID doesnt exist yet!! just create it :)
-                print(data['shares'])
-                data['shares'].append({
-                    "id": obj.getID(),
-                    "price": obj.getPrice()
-                })
 
-            data = str(data).replace("'", '"')
+                try: #to allow for multiple shares to be done you could like set a variable of percent change in the json, read that. if that *= 0.05 isn't less that the current percent change, don't buy again
+                    highestIDIndex = 0
 
-            file.write(data)
-            file.close()
+                    for i in range(0,len(data["shares"])): #iterate through all shares TRY ISNT GOING THROUGH
+                        if(data["shares"][i]["id"] == obj.getID()): # if the iteration of the shares has the same ID
+                            highestIDIndex = i
+                            print(highestIDIndex)
+
+                    if(float(data["shares"][highestIDIndex]["change"]) < (obj.getChange() - 0.01)): #it's checking for all shares. I need it to check the most recent one
+                            data["shares"].append({ # add new price to appropriate id
+                                "id": obj.getID(),
+                                "price": obj.getPrice(),
+                                "time": time.strftime("%H:%M:%S"),
+                                "change": obj.getChange()
+                            })
+                    else:
+                        print("Change not enough to buy another share!")
+
+                except AttributeError:
+                    # uh oh!!! that ID doesnt exist yet!! just create it :)
+                    data['shares'].append({
+                        "id": obj.getID(),
+                        "price": obj.getPrice(),
+                        "time": time.strftime("%H:%M:%S"),
+                        "change": obj.getChange()
+                    })
+
+                data = str(data).replace("'", '"')
+
+                file.write(data)
+                file.close()
 
     def sell(self, ID):
         obj = ShareObj(ID)
@@ -89,7 +105,7 @@ class ShareObj(object):
 
     def getChange(self):
         percent = self.share.get_percent_change()
-        return percent[:-1]
+        return float(percent[:-1])
 
     def getChangeFormatted(self):
         self.share.get_percent_change()
